MLE-634 Implemented review comments
